kv260_vit_project/requirements.txt
torch
torchvision
opencv-python
numpy
pillow
